requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
uuid==1.30
python-docx==0.8.11
python-magic==0.4.24
python-magic-bin==0.4.14
werkzeug==2.0.1 